pyproject.toml
[project]
name = "bookshelf-scanner"
version = "1.0.0"
description = "Scan bookshelves and export to Goodreads/StoryGraph"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "ultralytics>=8.0.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=10.0.0",
    "einops>=0.7.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
gpu = ["accelerate>=0.25.0"]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]

[project.scripts]
bookshelf-scanner = "bookshelf_scanner.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
